fix some bugs casued by oneflow

+++ utils/loss.py
@@ -196,7 +196,7 @@ class ComputeLoss:
 
         for i in range(self.nl):
             anchors, shape = self.anchors[i], p[i].shape
-            gain[2:6] = flow.tensor(shape)[[3, 2, 3, 2]]  # xyxy gain
+            gain[2:6] = flow.tensor(shape, dtype=gain.dtype)[[3, 2, 3, 2]]  # xyxy gain
 
             # Match targets to anchors
             t = targets * gain  # shape(3,n,7)
